Creating the components 'Cabecalho', 'CampoTexto', 'BarraLateral', and 'Banner,' and importing the project's images and fonts.
Continuing the development of the project, creating the 'Cabecalho' component using the project logo and the 'CampoTexto' component to represent the search field of the 'Cabecalho'. Importing project images and creating the 'BarraLateral' components, dynamically indicating which link is active with dynamic styles using Styled Components, and creating the 'Banner' component. Finally, importing the project fonts, using the CSS font-face feature in the 'EstilosGlobais' component, and utilizing reserved font files.