Creating the stylized components that represent the application's containers and creating the 'Galeria' and 'Titulo' components.
Creating the styled components in the 'App.jsx' file for 'AppContainer,' 'MainContainer,' and 'ConteudoGaleria,' which will serve as containers that encompass and define specific content within our application. Next, creating the 'Titulo' component that represents a standard application title, and the 'Galeria' component responsible for managing and displaying the application's images. Finally, creating the sub-components 'Tags', 'Populares', and 'Imagem' within 'Gallery' and using the 'tags.json' and 'fotos.json' files for dynamic rendering of the tags present in the images and the images themselves, respectively.